# doc/swaggerDef.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task Management System API",
    "version": "1.0.0",
    "description": "API documentation for the Task Management System. Includes authentication (Google OAuth), user, and task management endpoints.",
    "contact": {
      "name": "Your Name",
      "email": "your.email@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local development server"
    }
  ],
  "components": {
    "schemas": {
...
      "Task": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "status": {
            "type": "string",
            "enum": ["pending", "in-progress", "completed", "cancelled"]
          },
          "priority": { "type": "string", "enum": ["low", "medium", "high"] },
          "dueDate": { "type": "string", "format": "date-time" },
          "assignedTo": { "type": "string" },
          "createdBy": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "attachments": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "filename": { "type": "string" },
                "url": { "type": "string" },
                "size": { "type": "number" }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "googleOAuth": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://accounts.google.com/o/oauth2/auth",
            "scopes": {
              "profile": "Access to user profile",
              "email": "Access to user email"
            }
          }
        }
      }
    }
  },
  "security": [{ "googleOAuth": [] }],
  "paths": {
...
    "/dashboard": {
      "get": {
        "summary": "Get dashboard for logged-in user",
        "tags": ["Dashboard"],
        "security": [{ "googleOAuth": [] }],
        "responses": {
          "200": { "description": "Dashboard page" },
          "302": { "description": "Redirect to login if not authenticated" }
        }
      }
    },
...
    "/tasks": {
      "get": {
        "summary": "Get all tasks for the logged-in user",
        "tags": ["Task"],
        "security": [{ "googleOAuth": [] }],
        "responses": {
          "200": {
            "description": "List of tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Task" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new task",
        "tags": ["Task"],
        "security": [{ "googleOAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Task" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          }
        }
      }
    },
...
  }
}